added style to sign up page
