+++ client/src/App.js
@@ -13,7 +13,7 @@ import Home from './pages/Home';
 import Help from './pages/Help/Help';
 import Login from './pages/Login/Login';
 import Logout from './pages/Login/Logout';
-import Signup from './pages/Login/Signup';
+import Signup from './pages/Signup/Signup';
 import Cart from './pages/Cart/Cart';
 import Profile from './pages/Login/Profile';
