--- client/src/pages/Login/Signup.js
@@ -1,83 +1,0 @@
-import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
-import { useMutation } from '@apollo/client';
-import Auth from '../../utils/auth';
-import { ADD_USER } from '../../utils/mutations';
-
-function Signup(props) {
-    const [formState, setFormState] = useState({ email: '', password: '' });
-    const [addUser] = useMutation(ADD_USER);
-
-    const handleFormSubmit = async (event) => {
-        event.preventDefault();
-        const mutationResponse = await addUser({
-            variables: {
-                email: formState.email,
-                password: formState.password,
-                firstName: formState.firstName,
-                lastName: formState.lastName,
-            },
-        });
-        const token = mutationResponse.data.addUser.token;
-        Auth.login(token);
-    };
-
-    const handleChange = (event) => {
-        const { name, value } = event.target;
-        setFormState({
-            ...formState,
-            [name]: value,
-        });
-    };
-
-    return (
-        <div>
-            <h2>Signup</h2>
-            <form onSubmit={handleFormSubmit}>
-                <div>
-                    <input
-                        placeholder="First name"
-                        name="firstName"
-                        type="firstName"
-                        id="firstName"
-                        onChange={handleChange}
-                    />
-                </div>
-                <div>
-                    <input
-                        placeholder="Last name"
-                        name="lastName"
-                        type="lastName"
-                        id="lastName"
-                        onChange={handleChange}
-                    />
-                </div>
-                <div>
-                    <input
-                        placeholder="Email address"
-                        name="email"
-                        type="email"
-                        id="email"
-                        onChange={handleChange}
-                    />
-                </div>
-                <div>
-                    <input
-                        placeholder="Password"
-                        name="password"
-                        type="password"
-                        id="pwd"
-                        onChange={handleChange}
-                    />
-                </div>
-                <div>
-                    <button type="submit">Submit</button>
-                </div>
-            </form>
-            <h3>Already have an account? Go back to <Link to="/login">login</Link></h3>
-
-        </div>
-    );
-}
-
-export default Signup;
